requirements.txt
pytest==5.0.1
pytest-watch==4.2.0
mypy==0.780
black==19.10b0
pandas==1.2.3
redis==3.5.3
google-cloud-bigquery==2.11.0
